% Auto-generated: do not edit by hand
\name{idomDashComponent}

\alias{idomDashComponent}

\title{IdomDashComponent component}

\description{
IdomDashComponent allows integration between Dash and IDOM
}

\usage{
idomDashComponent(viewId=NULL)
}

\arguments{
\item{viewId}{Character. The view ID for this component instance}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

