% Auto-generated: do not edit by hand
\name{idomPlotlyDash}

\alias{idomPlotlyDash}

\title{IdomPlotlyDash component}

\description{

}

\usage{
idomPlotlyDash(viewId=NULL)
}

\arguments{
\item{viewId}{Character. The view ID for this component instance}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

