% Auto-generated: do not edit by hand
\name{reactpyDashComponent}

\alias{reactpyDashComponent}

\title{ReactPyDashComponent component}

\description{
ReactPyDashComponent allows integration between Dash and ReactPy
}

\usage{
reactpyDashComponent(viewId=NULL)
}

\arguments{
\item{viewId}{Character. The view ID for this component instance}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

