% Auto-generated: do not edit by hand
\name{idomDashComponent}

\alias{idomDashComponent}

\title{IdomDashComponent component}

\description{

}

\usage{
idomDashComponent(viewId=NULL, clientModuleUrl=NULL)
}

\arguments{
\item{viewId}{Character. The view ID for this component instance}

\item{clientModuleUrl}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

