% Auto-generated: do not edit by hand
\name{idomPlotlyDash}

\alias{idomPlotlyDash}

\title{IdomPlotlyDash component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
idomPlotlyDash(id=NULL, layoutUpdate=NULL, layoutEvent=NULL,
importSourceUrl=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{layoutUpdate}{Named list. An object describing a JSON patch that will update the layout's model}

\item{layoutEvent}{Named list. An event sent from the layout}

\item{importSourceUrl}{Character. A string defining the base URL where dynamically importable web modules can be found}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

